Drawing is now good.
Composition is complex but has been simplified from its peak.
I'm not happy about the timeouts in there but I can see how the
current algorithms can paint themselves into a corner.
We can try a backtracking algorithm later, although that will require
more changes

Likewise if drawing an outline times out, the form is removed.
That should still be fixed by fixing the outline algorithm though.

I still need to improve the layout code a little.

And layer pens might need to make a comeback.

Colouring is getting there.

It's simplified and constrained at the moment but ready to improve
now.

There are plenty of incidental bug fixes and utility code additions,
but they are all in the service of the above.